Refactor test to ensure assertion has the same types

+++ VendingMachineTests/CoinTests.cs
@@ -34,7 +34,7 @@ namespace VendingMachineTests
 
             machine.InsertCoin(dimeWeight, dimeDiameter);
 
-            Assert.Equal(0.10,machine.CurrentValue);
+            Assert.Equal((decimal) 0.10, machine.CurrentValue);
         }
 
     }
